moved fetches from game.js to HTTPConn in Nexus
