+++ code/nexus/WSClient.js
@@ -3,48 +3,48 @@ const config = require("./config.js")
 const socketURI = config.APIWSENDPOINT
 
 class WSClient {
-		constructor(board, writeTile) {
+    constructor(board, writeTile) {
 
-			// INITIALIZER DISPATCH
-			this.dispatch = {
-				"writeTile": writeTile,
-			}
+      // INITIALIZER DISPATCH
+      this.dispatch = {
+        "writeTile": writeTile,
+      }
 
-			this.board = board
+      this.board = board
 
-			// BIND CONTEXT 
-			this.handleMsg = this.handleMsg.bind(this)
+      // BIND CONTEXT 
+      this.handleMsg = this.handleMsg.bind(this)
 
-			// INITIALIZE WS CONNECTION & ATTACH LISTENERS
-			this.wsConn = new WebSocket(socketURI)
-			this.wsConn.on('open', () => console.log("CONNECTED"))
-			this.wsConn.on('message', this.handleMsg)
-			this.wsConn.on('close', () => console.log('Socket closed'))
-		}
+      // INITIALIZE WS CONNECTION & ATTACH LISTENERS
+      this.wsConn = new WebSocket(socketURI)
+      this.wsConn.on('open', () => console.log("CONNECTED"))
+      this.wsConn.on('message', this.handleMsg)
+      this.wsConn.on('close', () => console.log('Socket closed'))
+    }
 
-		convertBoard(){
-			let array = []
+    convertBoard(){
+      let array = []
 
-			for (var key in this.board) {
-				let coordinates = key.split("-")
-				let x = parseInt(coordinates[0], 10)
-				let y = parseInt(coordinates[1], 10)
-				array.push({x,y, color: this.board[key]})
-			}
-			return array
-		}
+      for (var key in this.board) {
+        let coordinates = key.split("-")
+        let x = parseInt(coordinates[0], 10)
+        let y = parseInt(coordinates[1], 10)
+        array.push({x,y, color: this.board[key]})
+      }
+      return array
+    }
 
-		handleMsg(msg){
-			let data = JSON.parse(msg)
-			console.log(data)
-			const {action, payload} = data
+    handleMsg(msg){
+      let data = JSON.parse(msg)
+      // console.log(data)
+      const {action, payload} = data
 
-			if (action && payload)
-				this.dispatch[action](payload)
+      if (action && payload)
+        this.dispatch[action](payload)
 
-		}
+    }
 
 
-	}
+  }
 
 module.exports = WSClient
