+++ code/nexus/config.js
@@ -1,7 +1,7 @@
 module.exports = {
-  HTTPPORT: 3000,
-  APIENDPOINT: 'http://theAPI.link',
-  APIWSENDPOINT: 'ws://theAPI.link/socket',
+  HTTPPORT: 3001,
+  APIENDPOINT: 'http://localhost:3000',
+  APIWSENDPOINT: 'ws://localhost:8080',
   INTERVAL: 250,
   BOARDDIMENSION: 500,
   GROUPID: 0,
