+++ code/nexus/index.js
@@ -2,9 +2,10 @@ const express = require('express')
 const app = express()
 const url = require('url');
 const bodyParser = require('body-parser');
+const fetch = require('node-fetch');
+
 const Game = require('./Game.js')
 const config = require("./config.js")
-const fetch = require('node-fetch');
 
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(bodyParser.json());
@@ -24,9 +25,10 @@ const game = new Game()
 // HELPERS
 function startInterval() {
   setInterval(() => {
-    console.log(queue)
+    // console.log(queue)
     if (queue.length > 0) {
       let nextPoint = queue.shift()
+      console.log("SENDING ", nextPoint)
       game.setTile(nextPoint.x, nextPoint.y, nextPoint.c)
     }
   }, config.INTERVAL)
@@ -34,7 +36,6 @@ function startInterval() {
 startInterval()
 
 function checkValidPoint(x,y) {
-  console.log(x, y)
   const integers = Number.isInteger(x) && Number.isInteger(y)
   const inRange = x >= 0 && y >= 0 && x < size && y < size 
   return integers && inRange
@@ -61,7 +62,6 @@ app.post('/set-tile', (req, res) => {
   const x = req.body.x
   const y = req.body.y
   const c = req.body.c
-  console.log(checkValidPoint(x,y))
   if (checkValidPoint(x,y)) {
     const coordinate = {x, y, c}
     queue.push(coordinate)
